move openAI to models folder
