+++ node-zerox/src/models/openAI.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
-import { CompletionArgs, CompletionResponse } from "./types";
-import { convertKeysToSnakeCase, encodeImageToBase64 } from "./utils";
+import { CompletionArgs, CompletionResponse } from "../types";
+import { convertKeysToSnakeCase, encodeImageToBase64 } from "../utils";
 
 export const getCompletion = async ({
   apiKey,
